Support adding multiple records to a recordset
Allow adding an entry to an existing record given by declaring the
modification as non-exclusive. This change makes it possible to create
records such as A, AAAA, MX, SRV, and TXT which are configured to
return more than one value in the query response.

The exclusive parameter defaults to 'yes' to preserve prior behavior
of the module.

Resolves #18